pyproject.toml
# SPDX-FileCopyrightText: Copyright (c) 2023-2026 Almaz Ilaletdinov <a.ilaletdinov@yandex.ru>
# SPDX-License-Identifier: MIT

[tool.poetry]
name = "revive-code-bot"
version = "0.1.0"
description = ""
authors = ["Almaz Ilaletdinov <a.ilaletdinov@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.14,<3.15"
gitpython = "3.1.45"
gunicorn = "23.0.0"
psycopg = { version = "3.3.2", extras = ["binary"] }
django = "6.0"
django-environ = "0.12.0"
pygithub = "2.8.1"
pyyaml = "6.0.3"
attrs = "25.4.0"
lxml = "6.0.2"
cron-validator = "1.0.8"
whitenoise = "6.11.0"
sentry-sdk = {extras = ["django"], version = "2.48.0"}
pika = "1.3.2"
django-extensions = "4.1"

[tool.poetry.group.dev.dependencies]
pytest = "9.0.2"
ruff = "0.14.9"
wemake-python-styleguide = "1.5.0"
refurb = "2.2.0"
pytest-cov = "7.0.0"
pytest-django = "4.11.1"
time-machine = "3.2.0"
model-bakery = "1.20.5"
faker = "39.0.0"
mypy = "1.19.1"
django-stubs = {extras = ["compatible-mypy"], version = "5.2.8"}
lxml-stubs = "0.5.1"
types-requests = "2.32.4.20250913"
requests-mock = "1.12.1"
pytest-randomly = "4.0.1"
ipython = "9.8.0"
types-pika = "1.2.0b1"
django-test-migrations = "^1.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
line_length = 119
skip = ["migrations/*.py", ".venv"]
multi_line_output = 3
include_trailing_comma = true

[tool.ruff]
target-version = "py311"  # TODO: up to 3.12
line-length = 120
preview = true
output-format = "concise"

[tool.ruff.lint]
select = ["ALL"]
preview = true
fixable = [
  "F401",   # Unused import
  "Q000",   # Quotes
  "COM812", # Trailing comma missing
  "I",
]
ignore = [
  "ARG001",  # Unused function argument
  "ARG002",  # Unused method argument
  "D106",    # Missing docstring in public nested class
  "D203",    # no-blank-line-before-class
  "D213",    # multi-line-summary-first-line
  "D401",    # First line of docstring should be in imperative mood
  "D418",    # Function decorated with `@overload` shouldn't contain a docstring
  "FBT001",  # Boolean-typed positional argument in function definition
  "FBT002",  # Boolean-typed positional argument in function definition
  "FIX002",  # Line contains TODO, consider resolving the issue
  "FLY002",  # We not use f-strings
  "RUF100",  # WPS primary linter
  "RUF001",  # Project contain cyrillic symbols
  "RUF002",  # Project contain cyrillic symbols
  "RET505",  # Unnecessary `elif` after `return` statement
  "RET506",  # Unnecessary `elif` after `raise` statement
  "UP030",   # We use explicit references
  "UP032",   # We not use f-strings
  "UP004",   # Class `PrayerReaction` inherits from `object`
  "TD",      # "todo" formats
  "PLR630",  # We not respect staticmethods "TCH003",  # Move standard library import `os.PathLike` into a type-checking block
  "DOC201",  # `return` is not documented in docstrin
  "DOC501",  # Raised exception `UnreachebleCaseError` missing from docstring
  "DOC402",  # `yield` is not documented in docstring
]

[tool.ruff.lint.flake8-quotes]
docstring-quotes = "double"
inline-quotes = "single"
multiline-quotes = "double"

[tool.ruff.lint.pylint]
max-args = 6

[tool.ruff.lint.per-file-ignores]
"src/tests/*" = [
  "S101",    # use of `assert` detected
  "INP001",  # File `...` is part of an implicit namespace package.
  "D",       # Docstrings
  "ANN",     # Annotations
  "PLR2004", # Magic value
  "PLR0913", # Too many arguments to function call
]
"*/migrations/*" = [
  "D100",
  "CPY001",
  "D101",
  "RUF012",
]

[tool.refurb]
ignore = [184]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
markers = [
  "integration: integration tests",
]

[tool.mypy]
plugins = ["mypy_django_plugin.main"]
mypy_path = "src"

[tool.django-stubs]
django_settings_module = "config.settings"
